requirements.txt
anthropic
llama-index-core
llama-index-llms-anthropic
llama-index-tools-tavily-research
llama-index-tools-weather
chainlit
python-dotenv
tavily-python
openai
tenacity
tiktoken
pydantic
pillow
pandas
matplotlib
fastapi
uvicorn
pyowm
